+++ flask_via/exceptions.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+
+"""
+flask_via.exceptions
+--------------------
+
+Custom exceptions which can be thrown by ``Flask-Via``.
+"""
+
+
+class ImproperlyConfigured(Exception):
+    """ Raised in the event ``Flask-Via`` has not been properly configured
+    """
+
+    pass